Split messages into several files.
