+++ Assets/Mirror/Runtime/GeneralTypeMessages.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d9e7f599ba3a945e4a5549c60c89da82
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
