+++ Assets/Mirror/Runtime/InternalMessages.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 041921a41e0b54ea489f4f992c539923
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
