+++ Assets/Mirror/Runtime/MessageBase.cs
@@ -1,0 +1,18 @@
+namespace Mirror
+{
+    public interface IMessageBase
+    {
+        void Deserialize(NetworkReader reader);
+
+        void Serialize(NetworkWriter writer);
+    }
+
+    public abstract class MessageBase : IMessageBase
+    {
+        // De-serialize the contents of the reader into this message
+        public virtual void Deserialize(NetworkReader reader) {}
+
+        // Serialize the contents of this message into the writer
+        public virtual void Serialize(NetworkWriter writer) {}
+    }
+}
