+++ Assets/Mirror/Runtime/MessageBase.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: a283e945921644607a54df6a918027df
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
