+++ Assets/Mirror/Runtime/PublicMessages.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 938f6f28a6c5b48a0bbd7782342d763b
+guid: 89838ad561f9541e1858c1fd99bb6525
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
