+++ Assets/Mirror/Runtime/RemoteCallMessage.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 3491002d8ee9d498d97979020c97738b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
